Tons of refactoring
